pyproject.toml
[project]
name = "data-science-nailgun"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyYAML",
    "scikit-learn",
    "xgboost",
    "pandas",
    "tqdm"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ds_nailgun*"]

[project.scripts]
post-experiment = "ds_nailgun.nailgun.post_experiment:main"
